chore: use more correct trace width for RF
For a two layer board the correct trace width for 50 Ohms should be
around 2,7 mm. This is quite thick and doesn't connect well to the
existing pads. Therefore it's reduced to 1,7 mm as mentioned in the
datasheet for the antenna.

Also a test point is added to connect analyzers directly to the RF
output.